Resource Management & Authentication Completed
Authentication subsystem  added (corresponding changes to the acccounts\models.py have already been effected)
Resource Management Subsystem completed. Most parts seem to be working and can pass the tests. There's a problem with the logic behind get teams and get machinery methods. This code is in need of review and shall be subject to change.