add the logic for resource management subsystem (#2)
* Update models.py

* Update tests.py

* Update tests.py

* Update models.py

* Update tests.py

* Update models.py

* Update models.py

* Update models.py

* Added new tests

Some of these tests do NOT pass. Requesting immediate review and help.

* Resource Management & Authentication Completed

Authentication subsystem  added (corresponding changes to the acccounts\models.py have already been effected)
Resource Management Subsystem completed. Most parts seem to be working and can pass the tests. There's a problem with the logic behind get teams and get machinery methods. This code is in need of review and shall be subject to change.

* Fixes as mentioned in the review

@SeyedParsa All of your reviews except those concerning the exceptions have been resolved.
I think we shouldn't touch the exceptions right now as that'd mess up with the tests.
Another thing that I couldn't resolve was the get_teams and get_machinery methods.

* Minor improvements as mentioned in reviews

* Changed the tests into more loose-fitting tests

* apply changes due to reviews and merge

* apply changes due to a review

* remove changes in User model

* remove related name for role

* minor changes due to review

* minor change due to review

Co-authored-by: SeyedParsa <seyedparsamirtaheri@gmail.com>